backend/requirements.txt
Django==4.2.3
psycopg2==2.9.6
factory-boy==3.3.0
djangorestframework==3.14.0
